aa_evan_airflow_tutorials/chapter09/requirements.txt
pandas
numpy
seaborn
matplotlib
requests
psycopg2-binary
streamlit
plotly
pytz
apache-airflow-providers-postgres==5.7.1